+++ moqt-core/src/lib.rs
@@ -1,5 +1,5 @@
 mod modules;
 pub use modules::moqt_client::MOQTClient;
+pub use modules::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
 pub use modules::send_stream_dispatcher_repository::SendStreamDispatcherRepository;
-pub use modules::track_namespace_manager_repository::TrackNamespaceManagerRepository;
 pub use modules::*;
