+++ moqt-core/src/modules.rs
@@ -2,8 +2,8 @@ pub mod constants;
 pub mod control_message_type;
 pub mod messages;
 pub mod moqt_client;
+pub mod pubsub_relation_manager_repository;
 pub mod send_stream_dispatcher_repository;
 pub mod subscription_models;
-pub mod track_namespace_manager_repository;
 pub mod variable_bytes;
 pub mod variable_integer;
