+++ moqt-core/src/modules/subscription_models.rs
@@ -1,3 +1,5 @@
-pub mod subscription_nodes;
+pub mod consumer_node;
+pub mod node_registory;
+pub mod producer_node;
 pub mod subscriptions;
 pub mod tracks;
