+++ moqt-server/src/lib.rs
@@ -3,12 +3,13 @@ use crate::modules::{
     buffer_manager,
     buffer_manager::{buffer_manager, BufferCommand},
     control_message_handler::*,
+    relation_manager::{commands::TrackCommand, manager::pubsub_relation_manager},
     send_stream_dispatcher::{send_stream_dispatcher, SendStreamDispatchCommand},
-    track_namespace_manager::{track_namespace_manager, TrackCommand},
 };
 use anyhow::{bail, Context, Ok, Result};
 use bytes::BytesMut;
 pub use moqt_core::constants;
+use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
 use moqt_core::{
     constants::{StreamDirection, UnderlayType},
     control_message_type::ControlMessageType,
@@ -19,7 +20,7 @@ use moqt_core::{
         moqt_payload::MOQTPayload,
     },
     variable_integer::write_variable_integer,
-    MOQTClient, TrackNamespaceManagerRepository,
+    MOQTClient,
 };
 use std::sync::Arc;
 use std::time::Duration;
@@ -114,7 +115,7 @@ impl MOQT {
         tokio::spawn(async move { buffer_manager(&mut buffer_rx).await });
 
         // Start track management thread
-        tokio::spawn(async move { track_namespace_manager(&mut track_namespace_rx).await });
+        tokio::spawn(async move { pubsub_relation_manager(&mut track_namespace_rx).await });
 
         // Start stream management thread
         tokio::spawn(async move { send_stream_dispatcher(&mut send_stream_rx).await });
@@ -279,9 +280,11 @@ async fn handle_connection(
     }
 
     // Delete pub/sub information related to the client
-    let track_namespace_manager =
-        modules::track_namespace_manager::TrackNamespaceManager::new(track_namespace_tx.clone());
-    let _ = track_namespace_manager.delete_client(stable_id).await;
+    let pubsub_relation_manager =
+        modules::relation_manager::interface::PubSubRelationManagerInterface::new(
+            track_namespace_tx.clone(),
+        );
+    let _ = pubsub_relation_manager.delete_client(stable_id).await;
 
     // Delete senders to the client
     send_stream_tx
@@ -324,8 +327,10 @@ async fn handle_incoming_bi_stream(
     let recv_stream = &mut stream.recv_stream;
     let shread_send_stream = &mut stream.shread_send_stream;
 
-    let mut track_namespace_manager =
-        modules::track_namespace_manager::TrackNamespaceManager::new(track_namespace_tx.clone());
+    let mut pubsub_relation_manager =
+        modules::relation_manager::interface::PubSubRelationManagerInterface::new(
+            track_namespace_tx.clone(),
+        );
     let mut send_stream_dispatcher =
         modules::send_stream_dispatcher::SendStreamDispatcher::new(send_stream_tx.clone());
 
@@ -348,7 +353,7 @@ async fn handle_incoming_bi_stream(
             &mut buf,
             UnderlayType::WebTransport,
             &mut client,
-            &mut track_namespace_manager,
+            &mut pubsub_relation_manager,
             &mut send_stream_dispatcher,
         )
         .await;
