+++ moqt-server/src/modules.rs
@@ -1,6 +1,6 @@
 pub mod buffer_manager;
 pub(crate) mod control_message_handler;
 pub(crate) mod handlers;
+pub(crate) mod relation_manager;
 pub(crate) mod send_stream_dispatcher;
 pub(crate) mod server_processes;
-pub(crate) mod track_namespace_manager;
