+++ moqt-server/src/modules/handlers/unsubscribe_handler.rs
@@ -1,9 +1,6 @@
 use anyhow::Result;
-
-use moqt_core::{
-    messages::control_messages::unsubscribe::Unsubscribe, MOQTClient,
-    TrackNamespaceManagerRepository,
-};
+use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
+use moqt_core::{messages::control_messages::unsubscribe::Unsubscribe, MOQTClient};
 
 // TODO: Not implemented yet
 #[allow(dead_code)]
@@ -17,7 +14,7 @@ pub(crate) enum UnSubscribeResponse {
 pub(crate) async fn _unsubscribe_handler(
     unsubscribe_message: Unsubscribe,
     _client: &mut MOQTClient, // TODO: Not implemented yet
-    _track_namespace_manager_repository: &mut dyn TrackNamespaceManagerRepository, // TODO: Not implemented yet
+    _pubsub_relation_manager_repository: &mut dyn PubSubRelationManagerRepository, // TODO: Not implemented yet
 ) -> Result<UnSubscribeResponse> {
     tracing::trace!("unsubscribe_handler start.");
