+++ moqt-server/src/modules/server_processes/subscribe_message.rs
@@ -1,19 +1,20 @@
 use crate::modules::handlers::subscribe_handler::subscribe_handler;
 use anyhow::{bail, Result};
 use bytes::BytesMut;
+use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
 use moqt_core::{
     messages::{
         control_messages::{subscribe::Subscribe, subscribe_ok::SubscribeOk},
         moqt_payload::MOQTPayload,
     },
-    MOQTClient, SendStreamDispatcherRepository, TrackNamespaceManagerRepository,
+    MOQTClient, SendStreamDispatcherRepository,
 };
 
 pub(crate) async fn process_subscribe_message(
     payload_buf: &mut BytesMut,
     client: &mut MOQTClient,
     write_buf: &mut BytesMut,
-    track_namespace_manager_repository: &mut dyn TrackNamespaceManagerRepository,
+    pubsub_relation_manager_repository: &mut dyn PubSubRelationManagerRepository,
     send_stream_dispatcher_repository: &mut dyn SendStreamDispatcherRepository,
 ) -> Result<Option<SubscribeOk>> {
     let subscribe_request_message = match Subscribe::depacketize(payload_buf) {
@@ -27,7 +28,7 @@ pub(crate) async fn process_subscribe_message(
     let result = subscribe_handler(
         subscribe_request_message,
         client,
-        track_namespace_manager_repository,
+        pubsub_relation_manager_repository,
         send_stream_dispatcher_repository,
     )
     .await;
