+++ moqt-server/src/modules/server_processes/subscribe_ok_message.rs
@@ -1,14 +1,15 @@
 use crate::modules::handlers::subscribe_ok_handler::subscribe_ok_handler;
 use anyhow::{bail, Result};
 use bytes::BytesMut;
+use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
 use moqt_core::{
     messages::{control_messages::subscribe_ok::SubscribeOk, moqt_payload::MOQTPayload},
-    MOQTClient, SendStreamDispatcherRepository, TrackNamespaceManagerRepository,
+    MOQTClient, SendStreamDispatcherRepository,
 };
 
 pub(crate) async fn process_subscribe_ok_message(
     payload_buf: &mut BytesMut,
-    track_namespace_manager_repository: &mut dyn TrackNamespaceManagerRepository,
+    pubsub_relation_manager_repository: &mut dyn PubSubRelationManagerRepository,
     send_stream_dispatcher_repository: &mut dyn SendStreamDispatcherRepository,
     client: &mut MOQTClient,
 ) -> Result<()> {
@@ -22,7 +23,7 @@ pub(crate) async fn process_subscribe_ok_message(
 
     subscribe_ok_handler(
         subscribe_ok_message,
-        track_namespace_manager_repository,
+        pubsub_relation_manager_repository,
         send_stream_dispatcher_repository,
         client,
     )
